rename isolated.js to routes-helpers.js
